harness(streaming): extract stream_llm_with_thinking helper
The astream_events loop that emits thinking.delta + thinking.completed
moves out of agents/synthesizer.py into a reusable agents/llm_streaming
module. Behavior unchanged for the canned synthesizer; the new helper
is what the agentic synthesizer and meta agent need next so all
dispatch modes can stream reasoning, not just canned.
